Bugfixes and motor calibration

+++ Drone/Drone/HelperFunctions.c
@@ -7,7 +7,7 @@
 
 #include "HelperFunctions.h"
 
-uint16_t map(uint16_t x, uint16_t in_min, uint16_t in_max, uint16_t out_min, uint16_t out_max)
+int32_t map(int32_t x, int32_t in_min, int32_t in_max, int32_t out_min, int32_t out_max)
 {
 	return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
 }
